SERVER-71606 Pass const int* instead of timeval* for LDAP_OPT_TIMELIMIT
GitOrigin-RevId: 23bf1977ab458221edebcc1574e3d267dfcdadfd

+++ src/mongo/db/modules/enterprise/src/ldap/ldap_options.cpp
@@ -52,7 +52,7 @@ MONGO_STARTUP_OPTIONS_STORE(LDAPOptions)(InitializerContext* context) {
 
     if (params.count("security.ldap.timeoutMS")) {
         globalLDAPParams->connectionTimeout =
-            Milliseconds(params["security.ldap.timeoutMS"].as<long>());
+            Milliseconds(params["security.ldap.timeoutMS"].as<int>());
     }
 
     if (params.count("security.ldap.retryCount")) {
